Create folder for calculation components
Create a calculations folder to house all the calculation components. Update imports to reflect changes.
