+++ client/src/components/calculations/ci.js
@@ -1,35 +1,36 @@
 import React, { useState, useEffect } from "react";
-import { onSubmit } from "../utils/calculationUtils";
-import "../styles/styles.css";
+import { onSubmit } from "../../utils/calculationUtils";
+import "../../styles/styles.css";
 
-export default function Addition({ patientObj }) {
-    const [val1, setVal1] = useState("");
-    const [val2, setVal2] = useState("");
-    const valueType = "Addition";
+export default function CardiacIndex({ patientObj }) {
+    const [co, setco] = useState("");
+    const [bsa, setbsa] = useState("");
+    const valueType = "Cardiac Index";
     const [calculatedValue, setCalculatedValue] = useState("");
     const [placeholderText, setPlaceholderText] = useState("");
 
     useEffect(() => {
-        setPlaceholderText((val1 === "" && val2 === "") ? "Enter calculation inputs" : "Missing inputs");
+        setPlaceholderText((co === "" && bsa === "") ? "Enter calculation inputs" : "Missing inputs");
+        setbsa((patientObj !== undefined) ? patientObj.bsa : bsa)
 
-        if (val1 === "" && val2 === "") {
+        if (co === "" || bsa === "") {
             setCalculatedValue("");
             return;
         }
 
-        const result = +(+val1 + +val2).toFixed(3);
+        const result = +(+co / +bsa).toFixed(3);
         setCalculatedValue(result);
-    }, [val1, val2]);
+    }, [co, bsa]);
 
     return (
         <div>
             <h1>{valueType}</h1>
             <form onSubmit={e => onSubmit(e, patientObj, { valueType: valueType, calculatedValue: calculatedValue })}>
                 <div>
-                    Value 1: <input name="value1" placeholder="Ex: 9.0" type="number" step="1.0" value={val1} onChange={e => setVal1(e.target.value)} />
+                    Cardiac Output (L/min): <input name="CO" placeholder="Ex: 4.3 L/min" type="number" value={co} onChange={e => setco(e.target.value)} />
                 </div>
                 <div>
-                    Value 2: <input name="value2" placeholder="Ex: 10.0" type="number" value={val2} onChange={e => setVal2(e.target.value)} />
+                    Body Surface Area (m<sup>2</sup>): <input name="BSA" placeholder="Ex: 1.9 m2" type="number" value={bsa} onChange={e => setbsa(e.target.value)} />
                 </div>
                 <div>
                     Output: <input type="text" placeholder={placeholderText} value={calculatedValue} readOnly />
