+++ client/src/components/calculations/papi.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
-import { onSubmit } from "../utils/calculationUtils";
-import "../styles/styles.css";
+import { onSubmit } from "../../utils/calculationUtils";
+import "../../styles/styles.css";
 
 export default function Papi({ patientObj }) {
     const [pasp, setPasp] = useState("");
