+++ client/src/components/calculations/transpulGradient.js
@@ -1,36 +1,35 @@
 import React, { useState, useEffect } from "react";
-import { onSubmit } from "../utils/calculationUtils";
-import "../styles/styles.css";
+import { onSubmit } from "../../utils/calculationUtils";
+import "../../styles/styles.css";
 
-export default function Dpg({ patientObj }) {
-    const [padp, setPadp] = useState("");
+export default function TranspulGradient({ patientObj }) {
+    const [map, setMap] = useState("");
     const [pcwp, setPcwp] = useState("");
-    const valueType = "Diastolic Pulmonary Gradient";
+    const valueType = "Transpulmonary Gradient";
     const [calculatedValue, setCalculatedValue] = useState("");
     const [placeholderText, setPlaceholderText] = useState("");
 
     useEffect(() => {
-        setPlaceholderText((padp === "" && pcwp === "") ? "Enter calculation inputs" : "Missing inputs");
+        setPlaceholderText((map === "" && pcwp === "") ? "Enter calculation inputs" : "Missing inputs");
 
-        if (padp === "" || pcwp === "") {
+        if (map === "" || pcwp === "") {
             setCalculatedValue("");
             return;
         }
 
-        const result = +(+padp - +pcwp).toFixed(3);
+        const result = +(+map - +pcwp).toFixed(3);
         setCalculatedValue(result);
-    }, [padp, pcwp]);
+    }, [map, pcwp]);
 
     return (
         <div>
             <h1>{valueType}</h1>
             <form onSubmit={e => onSubmit(e, patientObj, { valueType: valueType, calculatedValue: calculatedValue })}>
                 <div>
-                    Diastolic Pulmonary Artery Pressure (mmHg): <input name="padp" type="number" value={padp} onChange={e => setPadp(e.target.value)} />
-
+                    Mean Arterial Pressure (mmHg): <input name="MPA" placeholder="Ex: 72 mmHg" type="number" value={map} onChange={e => setMap(e.target.value)} />
                 </div>
                 <div>
-                    Pulmonary Capillary Wedge Pressure (mmHg): <input name="pcwp" type="number" value={pcwp} onChange={e => setPcwp(e.target.value)} />
+                    Pulmonary Capillary Wedge Pressure (mmHg): <input name="CVP" placeholder="Ex: 8 mmHg" type="number" value={pcwp} onChange={e => setPcwp(e.target.value)} />
                 </div>
                 <div>
                     Output: <input type="text" placeholder={placeholderText} value={calculatedValue} readOnly /> mmHg
